Cleaned up RandomSpaceSampler and fixed compilation error

+++ src/space_sampler/impl/RandomSpaceSampler.cpp
@@ -7,25 +7,25 @@ namespace next_best_view {
         SamplePointCloudPtr pointCloud = SamplePointCloudPtr(new SamplePointCloud());
         pointCloud->push_back(currentSpacePosition);
 
-        float width = mMapHelperPtr->getMetricWidth() * contractor;
-        float height = mMapHelperPtr->getMetricHeight() * contractor;
+        double width = mMapHelperPtr->getMetricWidth() * contractor;
+        double height = mMapHelperPtr->getMetricHeight() * contractor;
 
-        float xStart = std::max(0.f, currentSpacePosition[0] - (0.5f * width));
-        float xEnd = std::min(mMapHelperPtr->getMetricWidth(), currentSpacePosition[0] + (0.5f * width));
+        double xStart = std::max(0., currentSpacePosition[0] - (0.5 * width));
+        double xEnd = std::min((double) mMapHelperPtr->getMetricWidth(), currentSpacePosition[0] + (0.5 * width));
 
-        float yStart = std::max(0.f, currentSpacePosition[1] - (0.5f * height));
-        float yEnd = std::min(mMapHelperPtr->getMetricHeight(), currentSpacePosition[1] + (0.5f * height));
+        double yStart = std::max(0., currentSpacePosition[1] - (0.5 * height));
+        double yEnd = std::min((double) mMapHelperPtr->getMetricHeight(), currentSpacePosition[1] + (0.5 * height));
 
-        float xDist = xEnd - xStart;
-        float yDist = yEnd - yStart;
+        double xDist = xEnd - xStart;
+        double yDist = yEnd - yStart;
 
         while (pointCloud->size() < mSampleSize)
         {
             //random numbers between 0 and 1;
-            float xRandom = ((float) std::rand() / (RAND_MAX));
-            float yRandom = ((float) std::rand() / (RAND_MAX));
+            double xRandom = ((double) std::rand() / (RAND_MAX));
+            double yRandom = ((double) std::rand() / (RAND_MAX));
 
-            SimpleVector3 randomPoint(xStart + xRandom * xDist, yStart + yRandom * yDist, 0);
+            SimpleVector3 randomPoint(xStart + xRandom * xDist, yStart + yRandom * yDist, 0.);
 
             //make sure randomPoint does not intersect with an obstacle
             int8_t occupancyValue = mMapHelperPtr->getRaytracingMapOccupancyValue(randomPoint);
